refactor: split AwardCatalog.ts into per-category modules
- Created 11 category files (81-227 lines each)
- Extracted 6 utility functions to awardUtils.ts
- Converted AwardCatalog.ts to barrel re-export (47 lines)
- All files under FILE_MODULARITY_SPEC thresholds
- Zero consumer changes (barrel pattern)
